Move tracker files
