+++ src/entry.cpp
@@ -9,8 +9,8 @@
 #include "core/log_proofs.h"
 #include "core/settings.h"
 #include "core/shared.h"
-#include "core/trackers/arcdps_tracker.h"
-#include "core/trackers/unofficial_extras_tracker.h"
+#include "trackers/arcdps_tracker.h"
+#include "trackers/unofficial_extras_tracker.h"
 #include "providers/common/provider_registry.h"
 #include "providers/kpme/kpme_provider.h"
 #include "providers/wingman/wingman_provider.h"
