+++ src/trackers/arcdps_tracker.cpp
@@ -1,8 +1,8 @@
 #include "arcdps_tracker.h"
-#include "../log_proofs.h"
+#include "../core/log_proofs.h"
 
 bool ArcdpsTracker::IsAvailable() const {
-	return !LogProofs::unofficalExtrasEnabled;
+    return !LogProofs::unofficalExtrasEnabled;
 }
 
 void ArcdpsTracker::Initialize() {
@@ -12,17 +12,17 @@ void ArcdpsTracker::Shutdown() {
 }
 
 void ArcdpsTracker::OnPlayerJoin(const PlayerInfo& player) {
-	LogProofs::AddPlayerFromTracker(player);
+    LogProofs::AddPlayerFromTracker(player);
 }
 
 void ArcdpsTracker::OnPlayerLeave(const PlayerInfo& player) {
-	LogProofs::RemovePlayerFromTracker(player);
+    LogProofs::RemovePlayerFromTracker(player);
 }
 
 void ArcdpsTracker::OnSelfDetected(const PlayerInfo& self) {
-	LogProofs::SetSelfFromTracker(self);
+    LogProofs::SetSelfFromTracker(self);
 }
 
 void ArcdpsTracker::OnSquadClear() {
-	LogProofs::ClearPlayers();
+    LogProofs::ClearPlayers();
 }
