+++ src/trackers/unofficial_extras_tracker.cpp
@@ -1,8 +1,8 @@
 #include "unofficial_extras_tracker.h"
-#include "../log_proofs.h"
+#include "../core/log_proofs.h"
 
 bool UnofficialExtrasTracker::IsAvailable() const {
-	return LogProofs::unofficalExtrasEnabled;
+    return LogProofs::unofficalExtrasEnabled;
 }
 
 void UnofficialExtrasTracker::Initialize() {
@@ -12,17 +12,17 @@ void UnofficialExtrasTracker::Shutdown() {
 }
 
 void UnofficialExtrasTracker::OnPlayerJoin(const PlayerInfo& player) {
-	LogProofs::AddPlayerFromTracker(player);
+    LogProofs::AddPlayerFromTracker(player);
 }
 
 void UnofficialExtrasTracker::OnPlayerLeave(const PlayerInfo& player) {
-	LogProofs::RemovePlayerFromTracker(player);
+    LogProofs::RemovePlayerFromTracker(player);
 }
 
 void UnofficialExtrasTracker::OnSelfDetected(const PlayerInfo& self) {
-	LogProofs::SetSelfFromTracker(self);
+    LogProofs::SetSelfFromTracker(self);
 }
 
 void UnofficialExtrasTracker::OnSquadClear() {
-	LogProofs::ClearPlayers();
+    LogProofs::ClearPlayers();
 }
